Changed the name of the package session variable to com.wxparams.package.
The getForm factory method is now responsible for determining the
location of the form XML file.

Added a WxparamsConfig object for working with parameters. The object
specializes KObjectArray and provide a save method for storing changes
in the current configuration object. The object will create a default
configuration object if no configuration with the requested properties
(package and Itemid) is found. This object provide the default values
available in the form XML file.

A simple caching mechanism in the getConfig factory method was
implemented.